+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 **?.?.? Changes:**
 
+* General:
+  * Added config option to disable effect dispatching while run timer is paused
+
 * Twitch Voting:
   * Fixed vote options sometimes not being visible while dispatching is temporarily disabled
   * Improved error messages if the Twitch Client fails to connect